+++ sentry-android-fragment/src/main/java/io/sentry/android/fragment/FragmentLifecycleState.kt
@@ -1,0 +1,15 @@
+package io.sentry.android.fragment
+
+enum class FragmentLifecycleState(internal val breadcrumbName: String) {
+    ATTACHED("attached"),
+    SAVE_INSTANCE_STATE("save instance state"),
+    CREATED("created"),
+    VIEW_CREATED("view created"),
+    STARTED("started"),
+    RESUMED("resumed"),
+    PAUSED("paused"),
+    STOPPED("stopped"),
+    VIEW_DESTROYED("view destroyed"),
+    DESTROYED("destroyed"),
+    DETACHED("detached")
+}