+++ sentry-android-fragment/src/test/java/io/sentry/android/fragment/FragmentLifecycleIntegrationTest.kt
@@ -13,7 +13,7 @@ import com.nhaarman.mockitokotlin2.whenever
 import io.sentry.IHub
 import io.sentry.SentryOptions
 import kotlin.test.Test
-import kotlin.test.assertFalse
+import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
 class FragmentLifecycleIntegrationTest {
@@ -91,7 +91,7 @@ class FragmentLifecycleIntegrationTest {
             check { fragmentCallbacks ->
                 val callback = (fragmentCallbacks as SentryFragmentLifecycleCallbacks)
                 assertTrue(callback.enableAutoFragmentLifecycleTracing)
-                assertFalse(callback.enableFragmentLifecycleBreadcrumbs)
+                assertEquals(emptySet(), callback.filterFragmentLifecycleBreadcrumbs)
             },
             eq(true)
         )